dom/credentialmanagement/identity/tests/mochitest/test_two_providers.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Two Providers in a Credential.get()</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="head.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    SimpleTest.waitForExplicitFinish();
    setupTest("two_providers").then(
      function () {
        return navigator.credentials.get({
          identity: {
            providers: [{
              configURL: "https://example.net/tests/dom/credentialmanagement/identity/tests/mochitest/server_manifest.sjs",
              clientId: "mochitest",
              nonce: "nonce"
            },
            {
              configURL: "https://example.net/tests/dom/credentialmanagement/identity/tests/mochitest/server_manifest.sjs",
              clientId: "mochitest",
              nonce: "nonce2"
            }
          ]
          }
        });
      }
    ).then((cred) => {
      // This is not the long-term planned behavior, but is the current
      // state of the spec.
      // See https://fedidcg.github.io/FedCM/#dom-identitycredential-discoverfromexternalsource-slot
      // as of Sept 2022.
      ok(false, "incorrectly got a credential");
    }).catch((err) => {
      ok(true, "correctly got an error");
    }).finally(() => {
      SimpleTest.finish();
    })
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
